Clean up and optimise bashrc files
I removed the auto-completion alias script because it was very slow
(approx 0.2 seconds to load). The next big thing to sort out is nvm
which takes almost as long.

Before starting, it took approx 0.4 seconds to load everything, now it
takes about 0.2 seconds and I'm aiming to get it down to 0.05 seconds.