Remove custom tmux-get-window-name script
Replace it with direct manipulation of tmux's pane_current_path with the
tmux formatting stuff.

The new version has fewer "features" than my custom script. However,
because it's run directly in tmux and not as an external script, it
updates instantly on changing panes rather asynchronously when the
command returns.